Implement `Group` entities sync with remote state

+++ app/src/main/java/com/senex/timetable/domain/usecase/SyncAllGroups.kt
@@ -1,14 +1,23 @@
 package com.senex.timetable.domain.usecase
 
-class SyncAllGroups(
+import kotlinx.coroutines.flow.first
+import javax.inject.Inject
+
+class SyncAllGroups @Inject constructor(
     private val getAllRemoteGroups: GetAllRemoteGroups,
+    private val getAllGroups: GetAllGroups,
     private val deleteAllGroups: DeleteAllGroups,
     private val saveAllGroups: SaveAllGroups,
 ) {
     suspend operator fun invoke() {
-        getAllRemoteGroups()?.let {
-            deleteAllGroups()
-            saveAllGroups(it)
+        getAllRemoteGroups()?.let { remoteGroups ->
+            val localGroups = getAllGroups().first()
+            if (!localGroups.containsAll(remoteGroups)
+                || !remoteGroups.containsAll(localGroups)
+            ) {
+                deleteAllGroups()
+                saveAllGroups(remoteGroups)
+            }
         }
     }
 }
